split tests for `TrialOrIntroPriceEligibilityChecker` into SK1 and SK2
