+++ StoreKitUnitTests/TrialOrIntroPriceEligibilityCheckerSK2Tests.swift
@@ -7,7 +7,7 @@
 //
 //      https://opensource.org/licenses/MIT
 //
-//  TrialOrIntroPriceEligibilityCheckerTests.swift
+//  TrialOrIntroPriceEligibilityCheckerSK2Tests.swift
 //
 //  Created by César de la Vega on 9/1/21.
 
@@ -16,7 +16,8 @@ import Nimble
 import StoreKitTest
 import XCTest
 
-class TrialOrIntroPriceEligibilityCheckerTests: StoreKitConfigTestCase {
+// swiftlint:disable:next type_name
+class TrialOrIntroPriceEligibilityCheckerSK2Tests: StoreKitConfigTestCase {
 
     var receiptFetcher: MockReceiptFetcher!
     var trialOrIntroPriceEligibilityChecker: TrialOrIntroPriceEligibilityChecker!
@@ -25,29 +26,8 @@ class TrialOrIntroPriceEligibilityCheckerTests: StoreKitConfigTestCase {
     var mockProductsManager: MockProductsManager!
     var mockSystemInfo: MockSystemInfo!
 
-    func setupSK1() throws {
-        mockSystemInfo = try MockSystemInfo(platformFlavor: "xyz",
-                                            platformFlavorVersion: "123",
-                                            finishTransactions: true)
-        receiptFetcher = MockReceiptFetcher(requestFetcher: MockRequestFetcher(), systemInfo: mockSystemInfo)
-        let mockProductsManager = MockProductsManager(systemInfo: mockSystemInfo)
-        mockIntroEligibilityCalculator = MockIntroEligibilityCalculator(productsManager: mockProductsManager,
-                                                                        receiptParser: MockReceiptParser())
-        mockBackend = MockBackend()
-        let mockOperationDispatcher = MockOperationDispatcher()
-        let identityManager = MockIdentityManager(mockAppUserID: "app_user")
-        trialOrIntroPriceEligibilityChecker =
-        TrialOrIntroPriceEligibilityChecker(receiptFetcher: receiptFetcher,
-                                            introEligibilityCalculator: mockIntroEligibilityCalculator,
-                                            backend: mockBackend,
-                                            identityManager: identityManager,
-                                            operationDispatcher: mockOperationDispatcher,
-                                            productsManager: mockProductsManager)
-    }
-
-    func setUpSK2WithError() throws {
+    override func setUpWithError() throws {
         try super.setUpWithError()
-
         mockSystemInfo = try MockSystemInfo(platformFlavor: "xyz",
                                             platformFlavorVersion: "123",
                                             finishTransactions: true)
@@ -68,85 +48,6 @@ class TrialOrIntroPriceEligibilityCheckerTests: StoreKitConfigTestCase {
                                             productsManager: mockProductsManager)
     }
 
-    func testSK1CheckTrialOrIntroPriceEligibilityDoesntCrash() throws {
-        try setupSK1()
-        trialOrIntroPriceEligibilityChecker!.sk1CheckEligibility([]) { _ in
-        }
-    }
-
-    func testSK1CheckTrialOrIntroPriceEligibilityFetchesAReceipt() throws {
-        try setupSK1()
-        trialOrIntroPriceEligibilityChecker!.sk1CheckEligibility([]) { _ in
-        }
-
-        expect(self.receiptFetcher.receiptDataCalled) == true
-    }
-
-    func testSK1EligibilityIsCalculatedFromReceiptData() throws {
-        try setupSK1()
-        let stubbedEligibility = ["product_id": IntroEligibilityStatus.eligible]
-        mockIntroEligibilityCalculator.stubbedCheckTrialOrIntroductoryPriceEligibilityResult = (stubbedEligibility, nil)
-
-        var completionCalled = false
-        var maybeEligibilities: [String: IntroEligibility]?
-        trialOrIntroPriceEligibilityChecker!.sk1CheckEligibility([]) { (eligibilities) in
-            completionCalled = true
-            maybeEligibilities = eligibilities
-        }
-
-        expect(completionCalled).toEventually(beTrue())
-        let receivedEligibilities = try XCTUnwrap(maybeEligibilities)
-        expect(receivedEligibilities.count) == 1
-    }
-
-    func testSK1EligibilityIsFetchedFromBackendIfErrorCalculatingEligibility() throws {
-        try setupSK1()
-        let stubbedError = NSError(domain: RCPurchasesErrorCodeDomain,
-                                   code: ErrorCode.invalidAppUserIdError.rawValue,
-                                   userInfo: [:])
-        mockIntroEligibilityCalculator.stubbedCheckTrialOrIntroductoryPriceEligibilityResult = ([:], stubbedError)
-
-        let productId = "product_id"
-        let stubbedEligibility = [productId: IntroEligibility(eligibilityStatus: IntroEligibilityStatus.eligible)]
-        mockBackend.stubbedGetIntroEligibilityCompletionResult = (stubbedEligibility, nil)
-        var completionCalled = false
-        var maybeEligibilities: [String: IntroEligibility]?
-        trialOrIntroPriceEligibilityChecker!.sk1CheckEligibility([productId]) { (eligibilities) in
-            completionCalled = true
-            maybeEligibilities = eligibilities
-        }
-
-        expect(completionCalled).toEventually(beTrue())
-        let receivedEligibilities = try XCTUnwrap(maybeEligibilities)
-        expect(receivedEligibilities.count) == 1
-        expect(receivedEligibilities[productId]?.status) == IntroEligibilityStatus.eligible
-
-        expect(self.mockBackend.invokedGetIntroEligibilityCount) == 1
-    }
-
-    func testSK1ErrorFetchingFromBackendAfterErrorCalculatingEligibility() throws {
-        try setupSK1()
-        let productId = "product_id"
-
-        let stubbedError = NSError(domain: RCPurchasesErrorCodeDomain,
-                                   code: ErrorCode.invalidAppUserIdError.rawValue,
-                                   userInfo: [:])
-        mockIntroEligibilityCalculator.stubbedCheckTrialOrIntroductoryPriceEligibilityResult = ([:], stubbedError)
-
-        mockBackend.stubbedGetIntroEligibilityCompletionResult = ([:], stubbedError)
-        var completionCalled = false
-        var maybeEligibilities: [String: IntroEligibility]?
-        trialOrIntroPriceEligibilityChecker!.sk1CheckEligibility([productId]) { (eligibilities) in
-            completionCalled = true
-            maybeEligibilities = eligibilities
-        }
-
-        expect(completionCalled).toEventually(beTrue())
-        let receivedEligibilities = try XCTUnwrap(maybeEligibilities)
-        expect(receivedEligibilities.count) == 1
-        expect(receivedEligibilities[productId]?.status) == IntroEligibilityStatus.unknown
-    }
-
     // - Note: Xcode throws a warning about @available and #available being redundant, but they're actually necessary:
     // Although the method isn't supposed to be called because of our @available marks,
     // everything in this class will still be called by XCTest, and it will cause errors.
@@ -155,7 +56,6 @@ class TrialOrIntroPriceEligibilityCheckerTests: StoreKitConfigTestCase {
         guard #available(iOS 15.0, tvOS 15.0, macOS 12.0, watchOS 8.0, *) else {
             throw XCTSkip("Required API is not available for this test.")
         }
-        try setUpSK2WithError()
 
         let products = ["product_id",
                         "com.revenuecat.monthly_4.99.1_week_intro",
@@ -183,8 +83,6 @@ class TrialOrIntroPriceEligibilityCheckerTests: StoreKitConfigTestCase {
         guard #available(iOS 15.0, tvOS 15.0, macOS 12.0, watchOS 8.0, *) else {
             throw XCTSkip("Required API is not available for this test.")
         }
-        try setUpSK2WithError()
-
         let products = ["product_id",
                         "com.revenuecat.monthly_4.99.1_week_intro",
                         "com.revenuecat.annual_39.99.2_week_intro",
